Expand notification functionality into employees management

+++ API/Controllers/EmployeesController.cs
@@ -10,7 +10,8 @@ using Microsoft.AspNetCore.Mvc;
 namespace API.Controllers;
 
 [Authorize]
-public class EmployeesController(IUnitOfWork unitOfWork, IMapper mapper) : BaseApiController
+public class EmployeesController(IUnitOfWork unitOfWork, IMapper mapper,
+    INotificationService notificationService) : BaseApiController
 {
     [HttpGet]
     public async Task<ActionResult<PagedList<EmployeeDto>>> GetEmployees(
@@ -39,6 +40,13 @@ public class EmployeesController(IUnitOfWork unitOfWork, IMapper mapper) : BaseA
 
         unitOfWork.EmployeeRepository.AddEmployee(employee);
 
+        var department = await unitOfWork.DepartmentRepository.GetDepartmentDetailedByIdAsync(employee.DepartmentId);
+        if (department == null) return BadRequest("Failed to find department this employee belongs to");
+
+        notificationService.CreateNotifications(department.Managers, 
+            $"Department {department.Name} got new employee",
+            $"Department {department.Name} got new employee - {employee.FirstName} {employee.LastName}");
+
         if (await unitOfWork.Complete()) return Ok(mapper.Map<EmployeeDto>(employee));
         return BadRequest("Failed to create employee");
     }
@@ -49,6 +57,21 @@ public class EmployeesController(IUnitOfWork unitOfWork, IMapper mapper) : BaseA
     {
         var employee = await unitOfWork.EmployeeRepository.GetEmployeeByIdAsync(employeeId);
         if (employee == null) return BadRequest("Failed to find employee");
+
+        if (employee.DepartmentId != employeeEditDto.DepartmentId)
+        {
+            var previousDepartment = await unitOfWork.DepartmentRepository.GetDepartmentDetailedByIdAsync(employee.DepartmentId);
+            var currentDepartment = await unitOfWork.DepartmentRepository.GetDepartmentDetailedByIdAsync(employeeEditDto.DepartmentId);
+
+            if (previousDepartment == null || currentDepartment == null)
+                return BadRequest("Department does not exist");
+
+            var managers = previousDepartment.Managers.Concat(currentDepartment.Managers);
+            
+            notificationService.CreateNotifications(managers, 
+                $"Employee moved to new department",
+                $"Employee {employee.FirstName} {employee.LastName} has moved from {previousDepartment.Name} to {currentDepartment.Name}");
+        }
 
         mapper.Map(employeeEditDto, employee);
 
@@ -62,6 +85,13 @@ public class EmployeesController(IUnitOfWork unitOfWork, IMapper mapper) : BaseA
     {
         var employee = await unitOfWork.EmployeeRepository.GetEmployeeByIdAsync(employeeId);
         if (employee == null) return BadRequest("Failed to find employee");
+
+        var department = await unitOfWork.DepartmentRepository.GetDepartmentDetailedByIdAsync(employee.DepartmentId);
+        if (department == null) return BadRequest("Failed to find department this employee belongs to");
+
+        notificationService.CreateNotifications(department.Managers, 
+            $"Employee in department {department.Name} deleted",
+            $"Employee {employee.FirstName} {employee.LastName} has been deleted");
         
         unitOfWork.EmployeeRepository.DeleteEmployee(employee);
 
@@ -79,6 +109,13 @@ public class EmployeesController(IUnitOfWork unitOfWork, IMapper mapper) : BaseA
         if (employee.IsActive == true) return BadRequest("This employee is already activated");
         employee.IsActive = true;
 
+        var department = await unitOfWork.DepartmentRepository.GetDepartmentDetailedByIdAsync(employee.DepartmentId);
+        if (department == null) return BadRequest("Failed to find department this employee belongs to");
+
+        notificationService.CreateNotifications(department.Managers, 
+            $"Employee in department {department.Name} set as active",
+            $"Employee {employee.FirstName} {employee.LastName} has been set as active");
+
         if (await unitOfWork.Complete()) return Ok(mapper.Map<EmployeeDto>(employee));
         return BadRequest("Failed to active employee");
     }
@@ -92,6 +129,13 @@ public class EmployeesController(IUnitOfWork unitOfWork, IMapper mapper) : BaseA
 
         if (employee.IsActive == false) return BadRequest("This employee is not activated yet");
         employee.IsActive = false;
+
+        var department = await unitOfWork.DepartmentRepository.GetDepartmentDetailedByIdAsync(employee.DepartmentId);
+        if (department == null) return BadRequest("Failed to find department this employee belongs to");
+
+        notificationService.CreateNotifications(department.Managers, 
+            $"Employee in department {department.Name} set as not-active",
+            $"Employee {employee.FirstName} {employee.LastName} has been set as not-active");
 
         if (await unitOfWork.Complete()) return Ok(mapper.Map<EmployeeDto>(employee));
         return BadRequest("Failed to deactive employee");
